requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
cowsay==5.0
Django==4.1.5
django-ckeditor==6.5.1
django-colorfield==0.8.0
django-js-asset==2.0.0
djangorestframework==3.14.0
Faker==16.4.0
future==0.18.3
Pillow==9.4.0
python-dateutil==2.8.2
python-slugify==7.0.0
pytz==2022.7.1
serializers==0.2.4
six==1.16.0
sqlparse==0.4.3
text-unidecode==1.3
